+++ recipyCmd/recipycmd.py
@@ -10,6 +10,7 @@ from tinydb import where, Query
 
 from recipyCommon import config, utils
 from recipyCommon.config import read_config_file
+from recipyCommon.utils import get_run
 from recipyCommon.version_control import hash_file
 from recipyCmd.templating import render_run_template, render_debug_template
 
@@ -113,51 +114,8 @@ def gui(args):
     recipyGui.run(debug=args['--debug'], port=port)
 
 
-def get_run(db, id=None, latest=False, starts_with=False):
-    """Get the runs from db satisfying conditions passed as kwargs.
-    If run is unique returns a dict else list of dicts.
-
-    :param id: if provided will match the run id with this value.
-    :param latest: if True will return last run.
-    :param starts_with: if True and id exists will return all runs starting with id.
-    """
-    runs = None
-    if latest:
-        try:
-            runs = sorted(db.all(), key=lambda x: x['date'])[-1]
-        except (KeyError, IndexError) as _:
-            pass
-
-    elif starts_with and id:
-        runs = db.search(where('unique_id').matches('{}.*'.format(id)))
-        if len(runs) == 1:
-            runs = runs[0]
-
-    elif not starts_with and id:
-        try:
-            runs = db.search(where('unique_id') == id)[0]
-        except IndexError:
-            pass
-    return runs
-
-
-def get_latest_run():
-    results = db.all()
-
-    # If no runs in the database
-    if len(results) == 0:
-        return None
-
-    results = [_change_date(result) for result in results]
-
-    # Sort the results
-    results = sorted(results, key=lambda x: x['date'])
-
-    return results[-1]
-
-
 def latest(args):
-    run = get_latest_run()
+    run = get_run(db, latest=True)
 
     if not run:
         if args['--json']:
@@ -326,11 +284,3 @@ def search_text(args):
                         print(results[-1]['diff'])
 
     db.close()
-
-
-def _change_date(result):
-    result['date'] = str(result['date']).replace('{TinyDate}:', '')
-    return result
-
-if __name__ == '__main__':
-    main()
